fix edit and delete

+++ app/Http/Controllers/admin/FlatController.php
@@ -170,11 +170,11 @@ class FlatController extends Controller
     public function edit(Flat $flat)
     {
         //controllo che sia l'autore, se non lo è ridirigo sulla index
-        //  if($post->user_id !== Auth::id()){
-        //     return redirect()->route('admin.posts.index')
-        //     ->with('message', 'Non sei Autorizzato a modificare questo post')
-        //     ->with('type', 'warning');
-        // }
+          if($flat->user_id !== Auth::id()){
+             return redirect()->route('admin.flats.index')
+             ->with('message', 'Non sei Autorizzato a modificare questo appartament')
+             ->with('type', 'warning');
+         }
         $services = Service::select('id', 'label', 'icon')->get();
 
         $prev_services = $flat->services->pluck('id')->toArray();
@@ -216,12 +216,16 @@ class FlatController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Flat $flat)
-    {
+    {   
+        if ($flat->user_id !== Auth::id()) {
+            return redirect()->route('admin.flats.index')
+            ->with('message', "Non sei autorizzato ad eliminare questo appartamento")
+            ->with('type', "warning");
+        }
         $flat->services()->detach();
         $flat->messages()->delete();
 
         $flat->delete();
-
         return redirect()->route('admin.flats.index');
     }
